kernel, rinad: updated code to compile
