+++ include/irati/kucommon.h
@@ -677,6 +677,87 @@ struct media_report {
 	struct list_head available_difs;
 };
 
+struct name * rina_name_create(void);
+void rina_name_free(struct name *name);
+void flow_spec_free(struct flow_spec * fspec);
+struct flow_spec * rina_fspec_create(void);
+void policy_parm_free(struct policy_parm * prm);
+struct policy_parm * policy_parm_create(void);
+void policy_free(struct policy * policy);
+struct policy * policy_create(void);
+void dtp_config_free(struct dtp_config * dtp_config);
+struct dtp_config * dtp_config_create(void);
+void window_fctrl_config_free(struct window_fctrl_config * wfc);
+struct window_fctrl_config * window_fctrl_config_create(void);
+void rate_fctrl_config_free(struct rate_fctrl_config * rfc);
+struct rate_fctrl_config * rate_fctrl_config_create(void);
+void dtcp_fctrl_config_free(struct dtcp_fctrl_config * dfc);
+struct dtcp_fctrl_config * dtcp_fctrl_config_create(void);
+void dtcp_rxctrl_config_free(struct dtcp_rxctrl_config * rxfc);
+struct dtcp_rxctrl_config * dtcp_rxctrl_config_create(void);
+void dtcp_config_free(struct dtcp_config * dtcp_config);
+struct dtcp_config * dtcp_config_create(void);
+void pff_config_free(struct pff_config * pff);
+struct pff_config * pff_config_create(void);
+void rmt_config_free(struct rmt_config * rmt);
+struct rmt_config * rmt_config_create(void);
+void dt_cons_free(struct dt_cons * dtc);
+struct dt_cons * dt_cons_create(void);
+void qos_cube_free(struct qos_cube * qos);
+struct qos_cube * qos_cube_create(void);
+void efcp_config_free(struct efcp_config * efc);
+struct efcp_config * efcp_config_create(void);
+void fa_config_free(struct fa_config * fac);
+struct fa_config * fa_config_create(void);
+void resall_config_free(struct resall_config * resc);
+struct resall_config * resall_config_create(void);
+void et_config_free(struct et_config * etc);
+struct et_config * et_config_create(void);
+void static_ipcp_addr_free(struct static_ipcp_addr * addr);
+struct static_ipcp_addr * static_ipcp_addr_create(void);
+void address_pref_config_free(struct address_pref_config * apc);
+struct address_pref_config * address_pref_config_create(void);
+void addressing_config_free(struct addressing_config * ac);
+struct addressing_config * addressing_config_create(void);
+void nsm_config_free(struct nsm_config * nsmc);
+struct nsm_config * nsm_config_create(void);
+void auth_sdup_profile_free(struct auth_sdup_profile * asp);
+struct auth_sdup_profile * auth_sdup_profile_create(void);
+void secman_config_free(struct secman_config * sc);
+struct secman_config * secman_config_create(void);
+void routing_config_free(struct routing_config * rc);
+struct routing_config * routing_config_create(void);
+void ipcp_config_entry_free(struct ipcp_config_entry * ice);
+struct ipcp_config_entry * ipcp_config_entry_create(void);
+void dif_config_free(struct dif_config * dif_config);
+struct dif_config * dif_config_create(void);
+void rib_object_data_free(struct rib_object_data * rod);
+struct rib_object_data * rib_object_data_create(void);
+void query_rib_resp_free(struct query_rib_resp * qrr);
+struct query_rib_resp * query_rib_resp_create(void);
+void port_id_altlist_free(struct port_id_altlist * pia);
+struct port_id_altlist * port_id_altlist_create(void);
+void mod_pff_entry_free(struct mod_pff_entry * pffe);
+struct mod_pff_entry * mod_pff_entry_create(void);
+void pff_entry_list_free(struct pff_entry_list * pel);
+struct pff_entry_list * pff_entry_list_create(void);
+void sdup_crypto_state_free(struct sdup_crypto_state * scs);
+struct sdup_crypto_state * sdup_crypto_state_create(void);
+void dif_properties_entry_free(struct dif_properties_entry * dpe);
+struct dif_properties_entry * dif_properties_entry_create(void);
+void get_dif_prop_resp_free(struct get_dif_prop_resp * gdp);
+struct get_dif_prop_resp * get_dif_prop_resp_create(void);
+void ipcp_neighbor_free(struct ipcp_neighbor * nei);
+struct ipcp_neighbor * ipcp_neighbor_create(void);
+void ipcp_neigh_list_free(struct ipcp_neigh_list * nei);
+struct ipcp_neigh_list * ipcp_neigh_list_create(void);
+void bs_info_entry_free(struct bs_info_entry * bie);
+struct bs_info_entry * bs_info_entry_create(void);
+void media_dif_info_free(struct media_dif_info * mdi);
+struct media_dif_info * media_dif_info_create(void);
+void media_report_free(struct media_report * mre);
+struct media_report * media_report_create(void);
+
 #define IRATI_SUCC  0
 #define IRATI_ERR   1
