+++ kernel/efcp.c
@@ -278,7 +278,7 @@ int efcp_container_destroy(struct efcp_container * container)
                                                      efcp_destroy);
         if (container->cidm)       cidm_destroy(container->cidm);
 
-        if (container->config)     efcp_config_destroy(container->config);
+        if (container->config)     efcp_config_free(container->config);
 
 	if (container->rset)       rset_unregister(container->rset);
         rkfree(container);
