+++ kernel/policies.h
@@ -38,7 +38,6 @@ int                  policy_param_value_set(struct policy_parm * param,
 
 struct policy;
 
-struct policy *      policy_create(void);
 struct policy *      policy_dup_name_version(const struct policy * policy);
 struct policy *      policy_create_ni(void);
 int                  policy_destroy(struct policy * p);
