+++ kernel/sdup-crypto-ps-default.c
@@ -977,7 +977,7 @@ EXPORT_SYMBOL(default_sdup_update_crypto_state);
 
 struct ps_base * sdup_crypto_ps_default_create(struct rina_component * component)
 {
-	struct dup_config_entry * conf;
+	struct auth_sdup_profile * conf;
 	struct sdup_comp * sdup_comp;
 	struct sdup_crypto_ps * ps;
 	struct sdup_port * sdup_port;
@@ -1010,8 +1010,8 @@ struct ps_base * sdup_crypto_ps_default_create(struct rina_component * component
 	ps->dm          = sdup_port;
 	ps->priv        = data;
 
-	if (conf->crypto_policy) {
-		parameter = policy_param_find(conf->crypto_policy,
+	if (conf->encrypt) {
+		parameter = policy_param_find(conf->encrypt,
 					      "seq_win_size");
 		if (!parameter) {
 			data->seq_win_size = 0;
