+++ kernel/sdup-ttl-ps-default.c
@@ -114,7 +114,7 @@ EXPORT_SYMBOL(default_sdup_dec_check_lifetime_limit_policy);
 
 struct ps_base * sdup_ttl_ps_default_create(struct rina_component * component)
 {
-	struct dup_config_entry * conf;
+	struct auth_sdup_profile * conf;
 	struct sdup_comp * sdup_comp;
 	struct sdup_ttl_ps * ps;
 	struct sdup_port * sdup_port;
@@ -147,8 +147,8 @@ struct ps_base * sdup_ttl_ps_default_create(struct rina_component * component)
         ps->priv        = data;
 
         /* Parse parameters from config */
-	if (conf->ttl_policy) {
-		parameter = policy_param_find(conf->ttl_policy, "initialValue");
+	if (conf->ttl) {
+		parameter = policy_param_find(conf->ttl, "initialValue");
 		if (!parameter) {
 			LOG_ERR("Could not find 'initialValue' in TTL policy");
 			rkfree(ps);
