+++ librina/src/librina.pc.in
@@ -6,7 +6,7 @@ Name: LibRINA
 Description: RINA support libraries
 Version: @PACKAGE_VERSION@
 URL: @PACKAGE_URL@
-Requires: libnl-3.0 >= @LIBNL_MIN_VERSION@, libnl-genl-3.0 >= @LIBNL_MIN_VERSION@, protobuf >= @LIBPROTOBUF_MIN_VERSION@
+Requires: protobuf >= @LIBPROTOBUF_MIN_VERSION@
 Libs: -L${libdir} -lrina
-Libs.private: -lpthread @LIBNL3_LIBS@ @LIBNLGENL3_LIBS@ @LIBPROTOBUF_LIBS@
-Cflags: -I${includedir} @LIBNL3_CFLAGS@ @LIBNLGENL3_CFLAGS@ @LIBPROTOBUF_CFLAGS@
+Libs.private: -lpthread @LIBPROTOBUF_LIBS@
+Cflags: -I${includedir} @LIBPROTOBUF_CFLAGS@
