+++ rinad/src/ipcp/Makefile.am
@@ -28,6 +28,7 @@ COMMONCPPFLAGS =					\
 	-I$(srcdir)/../					\
 	-I$(srcdir)/../common				\
 	-I$(srcdir)/					\
+	-I$(top_srcdir)/../include			\
 	$(shimwifi_CPPFLAGS)
 
 COMMONLIBS =					\
